simplify codebase
